requirements.txt
mxnet-cu101
numpy
tqdm
tensorflow-cpu
gluoncv
gluoncv2
